requirements.txt
aiosqlite==0.17.0
anyio==3.7.1
APScheduler==3.10.4
asyncmy==0.2.8
asyncodbc==0.1.1
asyncpg==0.28.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iso8601==1.1.0
littlepaimon-utils==1.0.2
loguru==0.7.0
msgpack==1.0.5
multidict==6.0.4
nonebot-adapter-onebot==2.2.4
nonebot-plugin-apscheduler==0.4.0
nonebot-plugin-localstore==0.5.0
nonebot-plugin-tortoise-orm==0.1.1
nonebot2==2.2.1
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
Pillow==9.5.0
pydantic==1.10.12
pygtrie==2.5.0
pyodbc==4.0.39
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tomli==2.0.1
tortoise-orm==0.20.0
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0;sys_platform != 'win32'
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
